+++ projects/angular-jwt/src/lib/jwt.interceptor.ts
@@ -1,17 +1,23 @@
-import { Injectable, Inject } from "@angular/core";
+import { Injectable, Inject } from '@angular/core';
 import {
   HttpRequest,
   HttpHandler,
   HttpEvent,
   HttpInterceptor,
-} from "@angular/common/http";
-import { JwtHelperService } from "./jwthelper.service";
-import { JWT_OPTIONS } from "./jwtoptions.token";
+} from '@angular/common/http';
+import { DOCUMENT } from '@angular/common';
+import { JwtHelperService } from './jwthelper.service';
+import { JWT_OPTIONS } from './jwtoptions.token';
 
-import { mergeMap } from "rxjs/operators";
-import { parse } from "url";
-import { from, Observable } from "rxjs";
+import { map, mergeMap } from 'rxjs/operators';
+import { defer, from, Observable, of } from 'rxjs';
 
+const fromPromiseOrValue = <T>(input: T | Promise<T>) => {
+  if (input instanceof Promise) {
+    return defer(() => input);
+  }
+  return of(input);
+};
 @Injectable()
 export class JwtInterceptor implements HttpInterceptor {
   tokenGetter: (
@@ -19,43 +25,48 @@ export class JwtInterceptor implements HttpInterceptor {
   ) => string | null | Promise<string | null>;
   headerName: string;
   authScheme: string | ((request?: HttpRequest<any>) => string);
-  whitelistedDomains: Array<string | RegExp>;
-  blacklistedRoutes: Array<string | RegExp>;
+  allowedDomains: Array<string | RegExp>;
+  disallowedRoutes: Array<string | RegExp>;
   throwNoTokenError: boolean;
   skipWhenExpired: boolean;
-  standardPorts: string[] = ["80", "443"];
+  standardPorts: string[] = ['80', '443'];
 
   constructor(
     @Inject(JWT_OPTIONS) config: any,
-    public jwtHelper: JwtHelperService
+    public jwtHelper: JwtHelperService,
+    @Inject(DOCUMENT) private document: Document
   ) {
     this.tokenGetter = config.tokenGetter;
-    this.headerName = config.headerName || "Authorization";
+    this.headerName = config.headerName || 'Authorization';
     this.authScheme =
-      config.authScheme || config.authScheme === ""
+      config.authScheme || config.authScheme === ''
         ? config.authScheme
-        : "Bearer ";
-    this.whitelistedDomains = config.whitelistedDomains || [];
-    this.blacklistedRoutes = config.blacklistedRoutes || [];
+        : 'Bearer ';
+    this.allowedDomains = config.allowedDomains || [];
+    this.disallowedRoutes = config.disallowedRoutes || [];
     this.throwNoTokenError = config.throwNoTokenError || false;
     this.skipWhenExpired = config.skipWhenExpired;
   }
 
-  isWhitelistedDomain(request: HttpRequest<any>): boolean {
-    const requestUrl: any = parse(request.url, false, true);
-    const hostName =
-      requestUrl.hostname !== null
-        ? `${requestUrl.hostname}${
-            requestUrl.port && !this.standardPorts.includes(requestUrl.port)
-              ? ":" + requestUrl.port
-              : ""
-          }`
-        : requestUrl.hostname;
+  isAllowedDomain(request: HttpRequest<any>): boolean {
+    const requestUrl: URL = new URL(request.url, this.document.location.origin);
+
+    // If the host equals the current window origin,
+    // the domain is allowed by default
+    if (requestUrl.host === this.document.location.host) {
+      return true;
+    }
+
+    // If not the current domain, check the allowed list
+    const hostName = `${requestUrl.hostname}${
+      requestUrl.port && !this.standardPorts.includes(requestUrl.port)
+        ? ':' + requestUrl.port
+        : ''
+    }`;
 
     return (
-      hostName === null ||
-      this.whitelistedDomains.findIndex((domain) =>
-        typeof domain === "string"
+      this.allowedDomains.findIndex((domain) =>
+        typeof domain === 'string'
           ? domain === hostName
           : domain instanceof RegExp
           ? domain.test(hostName)
@@ -64,16 +75,22 @@ export class JwtInterceptor implements HttpInterceptor {
     );
   }
 
-  isBlacklistedRoute(request: HttpRequest<any>): boolean {
-    const requestedUrl = parse(request.url, false, true);
+  isDisallowedRoute(request: HttpRequest<any>): boolean {
+    const requestedUrl: URL = new URL(
+      request.url,
+      this.document.location.origin
+    );
 
     return (
-      this.blacklistedRoutes.findIndex((route: string | RegExp) => {
-        if (typeof route === "string") {
-          const parsedRoute = parse(route, false, true);
+      this.disallowedRoutes.findIndex((route: string | RegExp) => {
+        if (typeof route === 'string') {
+          const parsedRoute: URL = new URL(
+            route,
+            this.document.location.origin
+          );
           return (
             parsedRoute.hostname === requestedUrl.hostname &&
-            parsedRoute.path === requestedUrl.path
+            parsedRoute.pathname === requestedUrl.pathname
           );
         }
 
@@ -92,25 +109,32 @@ export class JwtInterceptor implements HttpInterceptor {
     next: HttpHandler
   ) {
     const authScheme = this.jwtHelper.getAuthScheme(this.authScheme, request);
-    let tokenIsExpired = false;
 
     if (!token && this.throwNoTokenError) {
-      throw new Error("Could not get token from tokenGetter function.");
+      throw new Error('Could not get token from tokenGetter function.');
     }
+
+    let tokenIsExpired = of(false);
 
     if (this.skipWhenExpired) {
-      tokenIsExpired = token ? this.jwtHelper.isTokenExpired(token) : true;
+      tokenIsExpired = token ? fromPromiseOrValue(this.jwtHelper.isTokenExpired(token)) : of(true);
     }
 
-    if (token && tokenIsExpired && this.skipWhenExpired) {
-      request = request.clone();
-    } else if (token) {
-      request = request.clone({
-        setHeaders: {
-          [this.headerName]: `${authScheme}${token}`,
-        },
-      });
+    if (token) {
+      return tokenIsExpired.pipe(
+        map((isExpired) =>
+          isExpired && this.skipWhenExpired
+            ? request.clone()
+            : request.clone({
+                setHeaders: {
+                  [this.headerName]: `${authScheme}${token}`,
+                },
+              })
+        ),
+        mergeMap((innerRequest) => next.handle(innerRequest))
+      );
     }
+
     return next.handle(request);
   }
 
@@ -118,22 +142,15 @@ export class JwtInterceptor implements HttpInterceptor {
     request: HttpRequest<any>,
     next: HttpHandler
   ): Observable<HttpEvent<any>> {
-    if (
-      !this.isWhitelistedDomain(request) ||
-      this.isBlacklistedRoute(request)
-    ) {
+    if (!this.isAllowedDomain(request) || this.isDisallowedRoute(request)) {
       return next.handle(request);
     }
     const token = this.tokenGetter(request);
 
-    if (token instanceof Promise) {
-      return from(token).pipe(
-        mergeMap((asyncToken: string | null) => {
-          return this.handleInterception(asyncToken, request, next);
-        })
-      );
-    } else {
-      return this.handleInterception(token, request, next);
-    }
+    return fromPromiseOrValue(token).pipe(
+      mergeMap((asyncToken: string | null) => {
+        return this.handleInterception(asyncToken, request, next);
+      })
+    );
   }
 }