+++ src/app/services/example-http.service.spec.ts
@@ -1,50 +1,54 @@
-import { TestBed } from "@angular/core/testing";
-import { ExampleHttpService } from "./example-http.service";
+import { fakeAsync, flush, TestBed } from '@angular/core/testing';
+import { ExampleHttpService } from './example-http.service';
 import {
   HttpClientTestingModule,
   HttpTestingController,
-} from "@angular/common/http/testing";
-import { JwtModule } from "angular-jwt";
+} from '@angular/common/http/testing';
+import { JwtModule } from 'angular-jwt';
 
 export function tokenGetter() {
-  return "TEST_TOKEN";
+  return 'TEST_TOKEN';
 }
 
 export function tokenGetterWithRequest(request) {
-  if (request.url.includes("1")) {
-    return "TEST_TOKEN_1";
+  if (request.url.includes('1')) {
+    return 'TEST_TOKEN_1';
   }
 
-  if (request.url.includes("2")) {
-    return "TEST_TOKEN_2";
+  if (request.url.includes('2')) {
+    return 'TEST_TOKEN_2';
   }
 
-  return "TEST_TOKEN";
+  return 'TEST_TOKEN';
 }
 
-describe("Example HttpService: with simple tokken getter", () => {
+export function tokenGetterWithPromise() {
+  return Promise.resolve('TEST_TOKEN');
+}
+
+describe('Example HttpService: with promise based tokken getter', () => {
   let service: ExampleHttpService;
   let httpMock: HttpTestingController;
 
   const validRoutes = [
     `/assets/example-resource.json`,
-    `http://whitelisted.com/api/`,
-    `http://whitelisted.com/api/test`,
-    `http://whitelisted.com:443/api/test`,
-    `http://whitelisted-regex.com/api/`,
-    `https://whitelisted-regex.com/api/`,
+    `http://allowed.com/api/`,
+    `http://allowed.com/api/test`,
+    `http://allowed.com:443/api/test`,
+    `http://allowed-regex.com/api/`,
+    `https://allowed-regex.com/api/`,
     `http://localhost:3000`,
     `http://localhost:3000/api`,
   ];
   const invalidRoutes = [
-    `http://whitelisted.com/api/blacklisted`,
-    `http://whitelisted.com/api/blacklisted-protocol`,
-    `http://whitelisted.com:80/api/blacklisted-protocol`,
-    `http://whitelisted.com/api/blacklisted-regex`,
-    `http://whitelisted-regex.com/api/blacklisted-regex`,
+    `http://allowed.com/api/disallowed`,
+    `http://allowed.com/api/disallowed-protocol`,
+    `http://allowed.com:80/api/disallowed-protocol`,
+    `http://allowed.com/api/disallowed-regex`,
+    `http://allowed-regex.com/api/disallowed-regex`,
     `http://foo.com/bar`,
-    "http://localhost/api",
-    "http://localhost:4000/api",
+    'http://localhost/api',
+    'http://localhost:4000/api',
   ];
 
   beforeEach(() => {
@@ -53,16 +57,12 @@ describe("Example HttpService: with simple tokken getter", () => {
         HttpClientTestingModule,
         JwtModule.forRoot({
           config: {
-            tokenGetter: tokenGetter,
-            whitelistedDomains: [
-              "whitelisted.com",
-              /whitelisted-regex*/,
-              "localhost:3000",
-            ],
-            blacklistedRoutes: [
-              "http://whitelisted.com/api/blacklisted-protocol",
-              "//whitelisted.com/api/blacklisted",
-              /blacklisted-regex*/,
+            tokenGetter: tokenGetterWithPromise,
+            allowedDomains: ['allowed.com', /allowed-regex*/, 'localhost:3000'],
+            disallowedRoutes: [
+              'http://allowed.com/api/disallowed-protocol',
+              '//allowed.com/api/disallowed',
+              /disallowed-regex*/,
             ],
           },
         }),
@@ -72,38 +72,117 @@ describe("Example HttpService: with simple tokken getter", () => {
     httpMock = TestBed.get(HttpTestingController);
   });
 
-  it("should add Authorisation header", () => {
+  it('should add Authorisation header', () => {
     expect(service).toBeTruthy();
   });
 
   validRoutes.forEach((route) =>
-    it(`should set the correct auth token for a whitelisted domain: ${route}`, () => {
+    it(`should set the correct auth token for a allowed domain: ${route}`, fakeAsync(() => {
+      service.testRequest(route).subscribe((response) => {
+        expect(response).toBeTruthy();
+      });
+
+      flush();
+      const httpRequest = httpMock.expectOne(route);
+
+      expect(httpRequest.request.headers.has('Authorization')).toEqual(true);
+      expect(httpRequest.request.headers.get('Authorization')).toEqual(
+        `Bearer TEST_TOKEN`
+      );
+    }))
+  );
+
+  invalidRoutes.forEach((route) =>
+    it(`should not set the auth token for a disallowed route: ${route}`, fakeAsync(() => {
+      service.testRequest(route).subscribe((response) => {
+        expect(response).toBeTruthy();
+      });
+
+      flush();
+      const httpRequest = httpMock.expectOne(route);
+      expect(httpRequest.request.headers.has('Authorization')).toEqual(false);
+    })
+  ));
+});
+
+describe('Example HttpService: with simple tokken getter', () => {
+  let service: ExampleHttpService;
+  let httpMock: HttpTestingController;
+
+  const validRoutes = [
+    `/assets/example-resource.json`,
+    `http://allowed.com/api/`,
+    `http://allowed.com/api/test`,
+    `http://allowed.com:443/api/test`,
+    `http://allowed-regex.com/api/`,
+    `https://allowed-regex.com/api/`,
+    `http://localhost:3000`,
+    `http://localhost:3000/api`,
+  ];
+  const invalidRoutes = [
+    `http://allowed.com/api/disallowed`,
+    `http://allowed.com/api/disallowed-protocol`,
+    `http://allowed.com:80/api/disallowed-protocol`,
+    `http://allowed.com/api/disallowed-regex`,
+    `http://allowed-regex.com/api/disallowed-regex`,
+    `http://foo.com/bar`,
+    'http://localhost/api',
+    'http://localhost:4000/api',
+  ];
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      imports: [
+        HttpClientTestingModule,
+        JwtModule.forRoot({
+          config: {
+            tokenGetter: tokenGetter,
+            allowedDomains: ['allowed.com', /allowed-regex*/, 'localhost:3000'],
+            disallowedRoutes: [
+              'http://allowed.com/api/disallowed-protocol',
+              '//allowed.com/api/disallowed',
+              /disallowed-regex*/,
+            ],
+          },
+        }),
+      ],
+    });
+    service = TestBed.get(ExampleHttpService);
+    httpMock = TestBed.get(HttpTestingController);
+  });
+
+  it('should add Authorisation header', () => {
+    expect(service).toBeTruthy();
+  });
+
+  validRoutes.forEach((route) =>
+    it(`should set the correct auth token for a allowed domain: ${route}`, () => {
       service.testRequest(route).subscribe((response) => {
         expect(response).toBeTruthy();
       });
 
       const httpRequest = httpMock.expectOne(route);
 
-      expect(httpRequest.request.headers.has("Authorization")).toEqual(true);
-      expect(httpRequest.request.headers.get("Authorization")).toEqual(
+      expect(httpRequest.request.headers.has('Authorization')).toEqual(true);
+      expect(httpRequest.request.headers.get('Authorization')).toEqual(
         `Bearer ${tokenGetter()}`
       );
     })
   );
 
   invalidRoutes.forEach((route) =>
-    it(`should not set the auth token for a blacklisted route: ${route}`, () => {
+    it(`should not set the auth token for a disallowed route: ${route}`, () => {
       service.testRequest(route).subscribe((response) => {
         expect(response).toBeTruthy();
       });
 
       const httpRequest = httpMock.expectOne(route);
-      expect(httpRequest.request.headers.has("Authorization")).toEqual(false);
+      expect(httpRequest.request.headers.has('Authorization')).toEqual(false);
     })
   );
 });
 
-describe("Example HttpService: with request based tokken getter", () => {
+describe('Example HttpService: with request based tokken getter', () => {
   let service: ExampleHttpService;
   let httpMock: HttpTestingController;
 
@@ -120,11 +199,7 @@ describe("Example HttpService: with request based tokken getter", () => {
         JwtModule.forRoot({
           config: {
             tokenGetter: tokenGetterWithRequest,
-            whitelistedDomains: [
-              "example-1.com",
-              "example-2.com",
-              "example-3.com",
-            ],
+            allowedDomains: ['example-1.com', 'example-2.com', 'example-3.com'],
           },
         }),
       ],
@@ -133,7 +208,7 @@ describe("Example HttpService: with request based tokken getter", () => {
     httpMock = TestBed.get(HttpTestingController);
   });
 
-  it("should add Authorisation header", () => {
+  it('should add Authorisation header', () => {
     expect(service).toBeTruthy();
   });
 
@@ -144,8 +219,8 @@ describe("Example HttpService: with request based tokken getter", () => {
       });
 
       const httpRequest = httpMock.expectOne(route);
-      expect(httpRequest.request.headers.has("Authorization")).toEqual(true);
-      expect(httpRequest.request.headers.get("Authorization")).toEqual(
+      expect(httpRequest.request.headers.has('Authorization')).toEqual(true);
+      expect(httpRequest.request.headers.get('Authorization')).toEqual(
         `Bearer ${tokenGetterWithRequest({ url: route })}`
       );
     })
@@ -153,9 +228,9 @@ describe("Example HttpService: with request based tokken getter", () => {
 });
 
 const authSchemes = [
-  [undefined, "Bearer "],
-  ["Basic ", "Basic "],
-  [() => "Basic ", "Basic "],
+  [undefined, 'Bearer '],
+  ['Basic ', 'Basic '],
+  [() => 'Basic ', 'Basic '],
 ];
 
 authSchemes.forEach((scheme) => {
@@ -163,9 +238,9 @@ authSchemes.forEach((scheme) => {
   let httpMock: HttpTestingController;
 
   describe(`Example HttpService: with ${
-    typeof scheme[0] === "function"
-      ? "an authscheme getter function"
-      : "a simple authscheme getter"
+    typeof scheme[0] === 'function'
+      ? 'an authscheme getter function'
+      : 'a simple authscheme getter'
   }`, () => {
     beforeEach(() => {
       TestBed.configureTestingModule({
@@ -175,7 +250,7 @@ authSchemes.forEach((scheme) => {
             config: {
               tokenGetter: tokenGetter,
               authScheme: scheme[0],
-              whitelistedDomains: ["whitelisted.com"],
+              allowedDomains: ['allowed.com'],
             },
           }),
         ],
@@ -185,13 +260,13 @@ authSchemes.forEach((scheme) => {
     });
 
     it(`should set the correct auth scheme a request (${scheme[1]})`, () => {
-      service.testRequest("http://whitelisted.com").subscribe((response) => {
+      service.testRequest('http://allowed.com').subscribe((response) => {
         expect(response).toBeTruthy();
       });
 
-      const httpRequest = httpMock.expectOne("http://whitelisted.com");
-      expect(httpRequest.request.headers.has("Authorization")).toEqual(true);
-      expect(httpRequest.request.headers.get("Authorization")).toEqual(
+      const httpRequest = httpMock.expectOne('http://allowed.com');
+      expect(httpRequest.request.headers.has('Authorization')).toEqual(true);
+      expect(httpRequest.request.headers.get('Authorization')).toEqual(
         `${scheme[1]}${tokenGetter()}`
       );
     });