+++ src/app/services/example-http.service.ts
@@ -1,13 +1,13 @@
-import { Injectable } from "@angular/core";
-import { HttpClient } from "@angular/common/http";
+import { Injectable } from '@angular/core';
+import { HttpClient } from '@angular/common/http';
 
 @Injectable({
-  providedIn: "root",
+  providedIn: 'root',
 })
 export class ExampleHttpService {
   constructor(private http: HttpClient) {}
 
-  testRequest(route = "/assets/example-resource.json") {
+  testRequest(route = '/assets/example-resource.json') {
     return this.http.get(route);
   }
 }